feat: Add right sidebar with following list
Introduces a new `RightSidebar` component to display a list of users the current user is following. This component will be positioned on the right side of the main content area, providing quick access to profiles of followed users. The main layout in `App.tsx` is updated to accommodate this new sidebar.